remove some majors and minors

+++ src/de/htwg/tictactoe/aview/gui/ModePlayerSwing.java
@@ -43,9 +43,9 @@ public class ModePlayerSwing extends JFrame implements ActionListener {
     private JButton reset;
     private JButton apply;    
     private ButtonGroup group;
-    private static float ROT = 20;
-    private static float GRUEN = 100;
-    private static float BLAU = 150;
+    private static int ROT = 20;
+    private static int GRUEN = 100;
+    private static int BLAU = 150;
     private int TEXTFIELDCOLUMN = 18;
     private int WIDTH = 700;
     private int HEIGHT = 250;
